pyproject.toml
[project]
name = "notificator_bot"
version = "1.0.0"
description = "A Telegram bot (aiogram) for announcements about new streams on Youtube or Twitch"
authors = [
    {name = "Alim",email = "zhaba.v.kedah77@gmail.com"}
]
requires-python = ">=3.13,<3.15"
dependencies = [
    "aiogram (>=3.26.0,<4.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "colorlog (>=6.10.1,<7.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false


[project.scripts]
start = "app"
